[MM-432] Fix issue of todo getting assigned to last assigned user
